requirements.txt
requests
beautifulsoup4
lxml
PyYAML
jinja2
feedparser
pdfminer.six
